Moving to tf lite
