+++ ciphey/basemods/Crackers/__init__.py
@@ -1,1 +1,1 @@
-from . import caesar
+from . import caesar, vigenere
