+++ ciphey/basemods/Crackers/caesar.py
@@ -18,12 +18,20 @@ import cipheydists
 from ciphey.iface import ParamSpec, CrackResults
 
 
-class Caesar(ciphey.iface.Cracker[str]):
+class Caesar(ciphey.iface.Cracker[str], ciphey.iface.Detector[str]):
     @staticmethod
     def getTargets() -> Set[str]:
         return {"caesar"}
 
-    def attemptCrack(self, message: str) -> Optional[CrackResults]:
+    def scoreLikelihood(self, ctext: str) -> Dict[str, float]:
+        # Match the distribution, and then run a chi-squared analysis
+        analysis = self.cache.get_or_update(ctext, "cipheycore::simple_analysis",
+                                            lambda: cipheycore.analyse_string(ctext))
+        return {"caesar": cipheycore.caesar_detect(analysis, self.expected)}
+
+    def attemptCrack(self, message: str, target: str) -> Optional[CrackResults]:
+        assert(target == "caesar")
+
         logger.debug("Trying caesar cipher")
         # Convert it to lower case
         #
@@ -86,4 +94,4 @@ class Caesar(ciphey.iface.Cracker[str]):
         return "caesar"
 
 
-ciphey.iface.registry.register(Caesar, ciphey.iface.Cracker[str])
+ciphey.iface.registry.register(Caesar, ciphey.iface.Cracker[str], ciphey.iface.Detector[str])
