+++ pyproject.toml
@@ -9,7 +9,6 @@ exclude = ["tests/hansard.txt"]
 
 [tool.poetry.dependencies]
 python = "^3.6"
-tensorflow = "^2.1.0"
 rich = "^1.2.3"
 loguru = "^0.5.0"
 pylint = "^2.5.2"
@@ -19,6 +18,7 @@ cipheycore = "^0.1.5"
 tflite-runtime = [
     {url = "https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp36-cp36m-linux_x86_64.whl", python="~3.6", markers="os_name == 'Linux'"},
     {url = "https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp37-cp37m-linux_x86_64.whl", python="~3.7", markers="os_name == 'Linux'"},
+    {url = "https://github.com/Ciphey/CipheyNeuralNetwork/raw/master/tflite/tflite_runtime-2.2.0-cp38-cp38-linux_x86_64.whl", python="~3.8", markers="os_name == 'Linux'"},
     {url = "https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp36-cp36m-macosx_10_14_x86_64.whl", python="~3.6", markers="os_name == 'Darwin'"},
     {url = "https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp37-cp37m-macosx_10_14_x86_64.whl", python="~3.7", markers="os_name == 'Darwin'"},
     {url = "https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp36-cp36m-win_amd64.whl", python="~3.6", markers="os_name == 'Windows'"},
